requirements/requirements.txt
--constraint "/constraints.txt"

apache-airflow-providers-snowflake==5.0.1
apache-airflow-providers-mysql==5.3.1
dbt-postgres==1.6.3
# we are adding astronomer-cosmos here
astronomer-cosmos==1.1.1
